% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_effects.R
\name{summarise_effects}
\alias{summarise_effects}
\title{Summarise Effects}
\usage{
summarise_effects(prior_data, current_data, summarise_type)
}
\arguments{
\item{prior_data}{The data.frame prior_data from inst/shiny_data}

\item{current_data}{The data.frame current_data from inst/shiny_data}

\item{summarise_type}{"order" for order-level summary (inst/shiny_data/meta_analysis_outputs.rds) or "all" for the overall effect on insects (inst/shiny_data/all-orders_meta_analysis_outputs.rds)}
}
\description{
Generates the datasets which are read by the Summary page of the shiny app. This script should be run twice every time new data is added, and the outputs saved to replace the previous outputs in inst/shiny_data. Setting summarise_type = "order" generates meta_analysis_outputs.rds; setting summarise_type = "all" generates all_orders_meta_analysis_outputs.rds.
}
